review views separated
