+++ ecommerce_project/myapp/views/__init__.py
@@ -2,3 +2,4 @@ from .user_views import *
 from .other_views import *
 from .product_related_views import *
 from .product_views import *
+from .review_views import *
