+++ ecommerce_project/myapp/views/product_related_views.py
@@ -9,26 +9,6 @@ from ecommerce_project.myapp.serializers.order_serializers import  \
     OrderLineOutputSerializer, OrderLineInputSerializer
 
 
-class ReviewListNCreateAPIView(APIView):
-    # permission_classes = [IsAuthenticated]
-    '''
-    only for list and creation
-    '''
-    def get(self, request, format=None):
-        review_list = Review.objects.all()
-        serializer = ReviewOutputSerializer(review_list, many=True)
-        return Response(serializer.data)
-
-    def post(self, request, format=None):
-        serializer = ReviewInputSerializer(data=request.data.copy())
-        if serializer.is_valid():
-            created_review = serializer.save()
-            print("I am here")
-            output_serializer = ReviewOutputSerializer(created_review)
-            return Response(output_serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
 class OrderListNCreateAPIView(APIView):
     # permission_classes = [IsAuthenticated]
     '''
